FEATURE: replace modifySuggestions with event
This adds an event after fetching suggestions.
A compatibility replacement is added to get the same behaviour as before.

The old implementation will generate deprecation warnings.
