+++ Classes/Controller/SearchController.php
@@ -7,7 +7,8 @@ use ApacheSolrForTypo3\Solr\System\Solr\SolrUnavailableException;
 use ApacheSolrForTypo3\Solr\Util;
 use Netlogix\Nxsolrajax\Domain\Search\ResultSet\SearchResultSet;
 use Netlogix\Nxsolrajax\Domain\Search\ResultSet\SuggestResultSet;
-use Netlogix\Nxsolrajax\SugesstionResultModifier;
+use Netlogix\Nxsolrajax\Event\Search\AfterGetSuggestionsEvent;
+use Psr\EventDispatcher\EventDispatcherInterface;
 use TYPO3\CMS\Core\Http\Response;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Frontend\Controller\TypoScriptFrontendController;
@@ -99,15 +100,12 @@ class SearchController extends \ApacheSolrForTypo3\Solr\Controller\SearchControl
             $searchRequest = $this->getSearchRequestBuilder()->buildForSuggest($arguments, $rawQuery, $pageId, $languageId);
             $result = $suggestService->getSuggestions($searchRequest, $additionalFilters);
 
-            if (is_array($GLOBALS['TYPO3_CONF_VARS']['EXTCONF']['nxsolrajax']['modifySuggestions'])) {
-                foreach ($GLOBALS['TYPO3_CONF_VARS']['EXTCONF']['nxsolrajax']['modifySuggestions'] as $key => $classRef) {
-                    $hookObject = GeneralUtility::makeInstance($classRef);
-                    if (!$hookObject instanceof SugesstionResultModifier) {
-                        throw new \Exception(sprintf('modifySuggestions hook expects SuggestionResultModifier, got %s', get_class($hookObject)), 1533224243);
-                    }
-                    $result['suggestions'] = $hookObject->modifySuggestions($queryString, is_array($result['suggestions']) ? $result['suggestions'] : [], $this->typoScriptConfiguration);
-                }
-            }
+
+            $event = new AfterGetSuggestionsEvent($queryString, $result['suggestions'] ?? [], $this->typoScriptConfiguration);
+            /** @var AfterGetSuggestionsEvent $event */
+            $event = GeneralUtility::makeInstance(EventDispatcherInterface::class)->dispatch($event);
+            $result['suggestions'] = $event->getSuggestions();
+
 
             $suggestResult = GeneralUtility::makeInstance(SuggestResultSet::class, $result['suggestions'], $result['suggestion']);
